refactoring code: form has been adjusted to pass values to the label and imput correctly without causing the app to crash.
grid had container adjusted in attempts to fix the padding in the fluid container. Possible solution is to do the fluid container with a row but fix has not been implemented.

navbar has had onclick events with arrow functions added to handle which modal is to open when clicked. Future updates should have logic to render the buttons if not signed in and not render if signed in.

MAJOR changes in modal. modal will now only act as a gateway for the form elements. all code within has been replaced with {props.children}. all form elements have been moved to the home page.

Changes on home:
State: state for username, password, firstname, and lastname have been added to track the user input for fields.

openModal: switch case has been added to differentiate between which modal to open on button click.

New functions:
handleFormSubmit: function added to handle the form submit and process an AJAX request. Logging is still in the document for development. Logging will need to be removed in a future release.

handleInputChange: function to track the user input on fields so the values can be submitted by the above function. Logging is still in the document for development. Logging will need to be removed in a furture release.

Render content:
modal body: modal body is now housed in a form inside of the called component. The nested components in SignInModal and SignUpModal are rendered by calling {props.children} in the modal component. All props for the form should be set and passed here.

TODOs added throughout home.

Api(client): basic api routes have been created to send the data from the form submits and to retrieve pictures and events. No rendering element is available for these get requests as of now.

App(routing): Routing has been completed for the current set of known pages. Commented code has been removed for a cleaner view. All rendered information should be written to the appropriate page.